+++ temperature.h
@@ -1,6 +1,6 @@
 #ifndef TEMPERATURE_H_ 
 #define TEMPERATURE_H_
 
-int temperature(void);  /* An example function declaration */
+uint16_t temperature(void);  /* An example function declaration */
 
-#endif // FOO_H_
+#endif
